+++ src/index/index.h
@@ -31,7 +31,7 @@ typedef struct {
 typedef struct {
 	char *name;
 	Attribute_ID id;
-	int64_t weight;
+	double weight;
 	bool nostem;
 	char *phonetic;
 } IndexField;
@@ -50,7 +50,7 @@ typedef struct {
 IndexField IndexField_New
 (
 	char *name,
-	int64_t weight,
+	double weight,
 	bool nostem,
 	char *phonetic
 );
